Migrate more classes and methods to specimen module
