+++ specimen/src/org/labkey/specimen/actions/SpecimenController.java
@@ -95,10 +95,9 @@ import org.labkey.api.specimen.model.SpecimenComment;
 import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
 import org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission;
 import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
-import org.labkey.api.specimen.settings.DisplaySettings;
-import org.labkey.api.specimen.settings.RepositorySettings;
-import org.labkey.api.specimen.settings.SettingsManager;
-import org.labkey.api.specimen.settings.StatusSettings;
+import org.labkey.specimen.settings.DisplaySettings;
+import org.labkey.specimen.settings.RepositorySettings;
+import org.labkey.specimen.settings.StatusSettings;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.Dataset;
 import org.labkey.api.study.SpecimenService;
@@ -178,6 +177,7 @@ import org.labkey.specimen.security.permissions.ManageRequestsPermission;
 import org.labkey.specimen.security.permissions.ManageSpecimenActorsPermission;
 import org.labkey.specimen.security.permissions.SetSpecimenCommentsPermission;
 import org.labkey.specimen.settings.RequestNotificationSettings;
+import org.labkey.specimen.settings.SettingsManager;
 import org.labkey.specimen.view.NotificationBean;
 import org.labkey.specimen.view.SpecimenRequestNotificationEmailTemplate;
 import org.labkey.specimen.view.SpecimenSearchWebPart;
