+++ specimen/src/org/labkey/specimen/query/SpecimenQueryView.java
@@ -52,9 +52,8 @@ import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
 import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
-import org.labkey.api.specimen.settings.DisplaySettings;
-import org.labkey.api.specimen.settings.RepositorySettings;
-import org.labkey.api.specimen.settings.SettingsManager;
+import org.labkey.specimen.settings.DisplaySettings;
+import org.labkey.specimen.settings.RepositorySettings;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
@@ -68,6 +67,7 @@ import org.labkey.api.view.DataView;
 import org.labkey.api.view.ViewContext;
 import org.labkey.api.writer.HtmlWriter;
 import org.labkey.specimen.actions.SpecimenController;
+import org.labkey.specimen.settings.SettingsManager;
 
 import java.io.IOException;
 import java.sql.ResultSet;
