+++ specimen/src/org/labkey/specimen/query/SpecimenRequestQueryView.java
@@ -30,7 +30,6 @@ import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.User;
 import org.labkey.api.specimen.SpecimenQuerySchema;
 import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
-import org.labkey.api.specimen.settings.SettingsManager;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.StudyUtils;
@@ -45,6 +44,7 @@ import org.labkey.specimen.actions.SpecimenController.DeleteRequestAction;
 import org.labkey.specimen.actions.SpecimenController.ManageRequestAction;
 import org.labkey.specimen.actions.SpecimenController.SubmitRequestAction;
 import org.labkey.specimen.security.permissions.ManageRequestsPermission;
+import org.labkey.specimen.settings.SettingsManager;
 
 import java.util.Set;
