+++ specimen/src/org/labkey/specimen/view/manageDisplay.jsp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 %>
-<%@ page import="org.labkey.api.specimen.settings.DisplaySettings"%>
+<%@ page import="org.labkey.specimen.settings.DisplaySettings"%>
 <%@ page import="org.labkey.api.study.StudyUrls"%>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
