+++ specimen/src/org/labkey/specimen/view/manageSpecimens.jsp
@@ -22,7 +22,6 @@
 <%@ page import="org.labkey.api.security.permissions.AdminPermission" %>
 <%@ page import="org.labkey.api.specimen.model.SpecimenTablesProvider" %>
 <%@ page import="org.labkey.api.specimen.security.permissions.ManageRequestSettingsPermission" %>
-<%@ page import="org.labkey.api.specimen.settings.SettingsManager" %>
 <%@ page import="org.labkey.api.study.SpecimenService" %>
 <%@ page import="org.labkey.api.study.SpecimenTransform" %>
 <%@ page import="org.labkey.api.study.Study" %>
@@ -42,6 +41,7 @@
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ManageSpecimenWebPartAction" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.ManageStatusesAction" %>
 <%@ page import="org.labkey.specimen.requirements.SpecimenRequestRequirementProvider" %>
+<%@ page import="org.labkey.specimen.settings.SettingsManager" %>
 <%@ page import="java.util.Collection" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
