+++ specimen/src/org/labkey/specimen/view/updateComments.jsp
@@ -17,9 +17,10 @@
 %>
 <%@ page import="org.apache.commons.lang3.StringUtils" %>
 <%@ page import="org.labkey.api.data.Container" %>
+<%@ page import="org.labkey.api.data.TableInfo" %>
 <%@ page import="org.labkey.api.security.User" %>
+<%@ page import="org.labkey.api.security.permissions.UpdatePermission" %>
 <%@ page import="org.labkey.api.specimen.Vial" %>
-<%@ page import="org.labkey.api.specimen.settings.SettingsManager" %>
 <%@ page import="org.labkey.api.study.Dataset" %>
 <%@ page import="org.labkey.api.study.Study" %>
 <%@ page import="org.labkey.api.study.StudyInternalService" %>
@@ -33,9 +34,8 @@
 <%@ page import="org.labkey.api.view.template.FrameFactoryClassic" %>
 <%@ page import="org.labkey.specimen.actions.SpecimenController.UpdateCommentsAction" %>
 <%@ page import="org.labkey.specimen.actions.UpdateSpecimenCommentsBean" %>
+<%@ page import="org.labkey.specimen.settings.SettingsManager" %>
 <%@ page import="java.util.Map" %>
-<%@ page import="org.labkey.api.data.TableInfo" %>
-<%@ page import="org.labkey.api.security.permissions.UpdatePermission" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%!
