+++ study/api-src/org/labkey/api/specimen/SpecimenMigrationService.java
@@ -10,7 +10,6 @@ import org.labkey.api.query.QuerySettings;
 import org.labkey.api.query.QueryUpdateService;
 import org.labkey.api.query.QueryView;
 import org.labkey.api.query.ValidationException;
-import org.labkey.api.reader.TabLoader;
 import org.labkey.api.security.User;
 import org.labkey.api.services.ServiceRegistry;
 import org.labkey.api.study.TimepointType;
@@ -18,7 +17,6 @@ import org.labkey.api.study.importer.SimpleStudyImportContext;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.ViewContext;
 
-import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Map;
@@ -27,14 +25,6 @@ import java.util.Map;
 // These should all go away once the migration is complete
 public interface SpecimenMigrationService
 {
-    // These constants were copied from SimpleSpecimenImporter, allowing that class to move into specimen
-    String VIAL_ID = "global_unique_specimen_id";
-    String SAMPLE_ID = "specimen_number";
-    String DRAW_TIMESTAMP = "draw_timestamp";
-    String VISIT = "visit_value";
-    String DERIVATIVE_TYPE = "derivative_type";
-    String PARTICIPANT_ID = "ptid";
-
     static SpecimenMigrationService get()
     {
         return ServiceRegistry.get().getService(SpecimenMigrationService.class);
@@ -58,11 +48,10 @@ public interface SpecimenMigrationService
 
     @Nullable QueryUpdateService getSpecimenQueryUpdateService(Container c, TableInfo queryTable);
 
-    void importSpecimens(Container container, User user, List<Map<String, Object>> specimens) throws ValidationException, IOException;
     void exportSpecimens(Container container, User user, List<Map<String, Object>> specimens, TimepointType timepointType, String participantIdLabel, HttpServletResponse response);
-    Map<String, String> getColumnLabelMap(Container container, User user);
-    void fixupSpecimenColumns(Container container, User user, TabLoader loader) throws IOException;
     QueryView getSpecimenQueryView(ViewContext context, QuerySettings settings);
 
     void setDefaultRequestabilityRules(Container container, User user);
+
+    boolean isEnableRequests(Container c);
 }
